# components/onboarding/api/swagger.yaml
basePath: /
definitions:
...
  CreateSegmentInput:
    description: CreateSegmentInput is the input payload to create a segment within
      a ledger, representing a logical division such as a business area, product line,
      or customer category.
    properties:
      metadata:
        additionalProperties: {}
        description: |-
          Additional custom attributes for the segment
          Keys max length: 100 characters, Values max length: 2000 characters
        type: object
      name:
        description: Name of the segment (required, max length 256 characters)
        example: My Segment
        maxLength: 256
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Status of the segment (active, inactive, pending)
    required:
    - name
    type: object
  Error:
    description: Standardized error response format used across all API endpoints
      for error situations. Provides structured information about errors including
      codes, messages, and field-specific validation details.
    properties:
      code:
        description: |-
          Error code identifying the specific error condition
          example: ERR_INVALID_INPUT
          maxLength: 50
        example: ERR_INVALID_INPUT
        maxLength: 50
        type: string
      entityType:
        description: |-
          Optional type of entity associated with the error
          example: Organization
          maxLength: 100
        example: Organization
        maxLength: 100
        type: string
      fields:
        additionalProperties:
          type: string
        description: |-
          Optional detailed field validations for client-side handling
          example: {"name": "Field 'name' is required"}
        example:
          '{"name"': ' "Field ''name'' is required"}'
        type: object
      message:
        description: |-
          Detailed error message explaining the issue
          example: The request contains invalid fields. Please check the field 'name' and try again.
          maxLength: 500
        example: The request contains invalid fields. Please check the field 'name'
          and try again.
        maxLength: 500
        type: string
      title:
        description: |-
          Short, human-readable error title
          example: Bad Request
          maxLength: 100
        example: Bad Request
        maxLength: 100
        type: string
    type: object
...
  Pagination:
    description: Pagination is the struct designed to store the pagination data of
      an entity list.
    properties:
      items: {}
      limit:
        example: 10
        type: integer
      next_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
      page:
        example: 1
        type: integer
      prev_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
    type: object
...
  Segment:
    description: Segment represents a logical division within a ledger such as a business
      area, product line, or customer category.
    properties:
      createdAt:
        description: Timestamp when the segment was created
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      deletedAt:
        description: Timestamp when the segment was deleted (null if not deleted)
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      id:
        description: Unique identifier for the segment (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      ledgerId:
        description: ID of the ledger this segment belongs to (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      metadata:
        additionalProperties: {}
        description: Additional custom attributes for the segment
        type: object
      name:
        description: Name of the segment (max length 256 characters)
        example: My Segment
        maxLength: 256
        type: string
      organizationId:
        description: ID of the organization that owns this segment (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Status of the segment (active, inactive, pending)
      updatedAt:
        description: Timestamp when the segment was last updated
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
  Status:
    description: 'Entity status information with a standardized code and optional
      description. Common status codes include: ACTIVE, INACTIVE, PENDING, SUSPENDED,
      DELETED.'
    properties:
      code:
        description: 'Status code identifier, common values include: ACTIVE, INACTIVE,
          PENDING, SUSPENDED, DELETED'
        example: ACTIVE
        maxLength: 100
        type: string
      description:
        description: Optional human-readable description of the status
        example: Active status
        maxLength: 256
        type: string
    type: object
...
  UpdateSegmentInput:
    description: UpdateSegmentInput is the input payload to update an existing segment's
      properties such as name, status, and metadata.
    properties:
      metadata:
        additionalProperties: {}
        description: |-
          Updated or additional custom attributes for the segment
          Keys max length: 100 characters, Values max length: 2000 characters
        type: object
      name:
        description: Updated name of the segment (optional, max length 256 characters)
        example: My Segment Updated
        maxLength: 256
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Updated status of the segment (active, inactive, pending)
    type: object
host: localhost:3000
info:
  contact:
    name: Discord community
    url: https://discord.gg/DnhqKwkGv3
  description: This is a swagger documentation for the Midaz Ledger API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Midaz Onboarding API
  version: v1.48.0
paths:
...
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/segments:
    get:
      description: Returns a paginated list of segments within the specified ledger,
        optionally filtered by metadata, date range, and other criteria
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: JSON string to filter segments by metadata fields
        in: query
        name: metadata
        type: string
      - default: 10
        description: Maximum number of records to return per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Filter segments created on or after this date (format: YYYY-MM-DD)'
        in: query
        name: start_date
        type: string
      - description: 'Filter segments created on or before this date (format: YYYY-MM-DD)'
        in: query
        name: end_date
        type: string
      - description: Sort direction for results based on creation date
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved segments list
          schema:
            allOf:
            - $ref: '#/definitions/Pagination'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/Segment'
                  type: array
                limit:
                  type: integer
                page:
                  type: integer
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization or ledger not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: List all segments
      tags:
      - Segments
    post:
      consumes:
      - application/json
      description: Creates a new segment within the specified ledger. Segments represent
        logical divisions within a ledger, such as business areas, product lines,
        or customer categories.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Segment details including name, status, and optional metadata
        in: body
        name: segment
        required: true
        schema:
          $ref: '#/definitions/CreateSegmentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created segment
          schema:
            $ref: '#/definitions/Segment'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization or ledger not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Segment with the same name already exists'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Create a new segment
      tags:
      - Segments
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/segments/{id}:
    delete:
      description: Permanently removes a segment from the specified ledger. This operation
        cannot be undone.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Segment ID in UUID format
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Segment successfully deleted
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Segment, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Segment cannot be deleted due to existing dependencies'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Delete a segment
      tags:
      - Segments
    get:
      description: Returns detailed information about a segment identified by its
        UUID within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Segment ID in UUID format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved segment
          schema:
            $ref: '#/definitions/Segment'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Segment, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Retrieve a specific segment
      tags:
      - Segments
    patch:
      consumes:
      - application/json
      description: Updates an existing segment's properties such as name, status,
        and metadata within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Segment ID in UUID format
        in: path
        name: id
        required: true
        type: string
      - description: Segment properties to update including name, status, and optional
          metadata
        in: body
        name: segment
        required: true
        schema:
          $ref: '#/definitions/UpdateSegmentInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated segment
          schema:
            $ref: '#/definitions/Segment'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Segment, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Segment with the same name already exists'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Update a segment
      tags:
      - Segments
...
swagger: "2.0"